バージョン表示、敬称、区分等 #35

+++ guest_types.php
@@ -2,12 +2,10 @@
 require_once("inc/class.dbo.php");
 include_once("inc/checklogin.inc.php");
 $obj = new DBO();
-$post = $obj->protectXSS($_POST);
 include_once("inc/header.inc.php");
-$get = $obj->protectXSS($_GET);
+
+	$get = $obj->protectXSS($_GET);
 	$post = $obj->protectXSS($_POST);
-
-
 
 	if(isset($post['name']) && $post['name'] != '')
 	{
@@ -34,27 +32,36 @@ $get = $obj->protectXSS($_GET);
 		}
 
 	}
-	if($_GET['action']=='sort' && (int)$_GET['id'] > 0) // UCHIDA EDIT 11/08/22
+	if(isset($get['action']) && $get['action']=='sort' && (int)$get['id'] > 0)
 	{
 		$table = 'spssp_guest_type';
 
-		$id = (int)$_GET['id'];
-		$move = $_GET['move'];
+		$id = (int)$get['id'];
+		$move = $get['move'];
 		$obj->sortItem($table, $id, $move);
+		$edit_id=$get['edit_id'];
+		$rTitle=$get['title'];
 	}
-	else if(isset($get['action']) && $get['action'] !== '')
+	else if(isset($get['action']) && $get['action'] == 'delete' && (int)$get['id'] > 0)
 	{
 		$id = (int)$get['id'];
 		if($id > 0)
 		{
 			$obj->DeleteRow("spssp_guest_type", "id=".$id);
 			$msg = 3;
+			if ((int)$get['id'] != (int)$get['edit_id']) $edit_id=$get['edit_id'];
+			$rTitle=$get['title'];
 		}
 		else
 		{
 			$err = 11;
 		}
 	}
+	if (isset($get['action']) && $get['action'] == 'edit' && (int)$get['id'] > 0) {
+		$edit_id=$get['edit_id'];
+		$rTitle=$get['title'];
+	}
+
 ?>
  <script type="text/javascript">
  var kindArray=new Array(); // UCHIDA EDIT 11/08/22
@@ -72,9 +79,9 @@ $get = $obj->protectXSS($_GET);
 	    <div  id="new_table" style="width:100%;">
 <!--         	<h2 id="title_bar">登録　区分: </h2> -->
 			<div class="subtitle">登録　区分: </div>
-        	<form action="guest_types.php" method="post" name="new_name">
-            	<input type="hidden" name="insert_edit" id="insert_edit" value="0" />
-                	区分 : &nbsp;<input type="text" name="name" id="name"  /> &nbsp; &nbsp; &nbsp;
+        	<form action="guest_types.php" method="post" name="new_name" >
+            	<input type="hidden" name="insert_edit" id="insert_edit" value="<?=$edit_id?>" />
+                	区分 : &nbsp;<input type="text" name="name" id="name" value="<?=$rTitle?>"/> &nbsp; &nbsp; &nbsp;
                 <a onclick="check_name()" href="javascript:void(0)">
 						<img src="img/common/btn_regist_update_admin.jpg"" alt="登録">
 				</a>
@@ -119,22 +126,22 @@ $get = $obj->protectXSS($_GET);
 			<script language="javascript" type="text/javascript">
 			kindArray[<?=$i?>]="<?=$row['name']?>";
 			</script>
-			<div class="<?=$class?>">
+			<div class="<?=$class?>" id="boxid<?=$row['id']?>">
             		<table border="0" align="center" cellpadding="1" cellspacing="1">
             			<tr align="center">
                         	<td width="10%"><?=$j?></td>
                             <td><?=$row['name']?></td>
 							<td>
-							<span class="txt1"><a href="guest_types.php?action=sort&amp;move=up&amp;id=<?=$row['id']?>">▲</a>
-							<a href="guest_types.php?action=sort&amp;move=down&amp;id=<?=$row['id']?>"> ▼</a></span>
+							<span class="txt1"><a href="guest_types.php?action=sort&amp;move=up&amp;id=<?=$row['id']?>&title=<?=$rTitle?>&edit_id=<?=$edit_id?>">▲</a>
+							<a href="guest_types.php?action=sort&amp;move=down&amp;id=<?=$row['id']?>&title=<?=$rTitle?>&edit_id=<?=$edit_id?>"> ▼</a></span>
 							</td>
                             <td>
-                            	<a href="#" onclick="edit_name(<?=$row['id']?>,'<?=$row['name']?>');">
+                            	<a href="javascript:void(0);" onclick="edit_name(<?=$row['id']?>,'<?=$row['name']?>');">
                                 	<img src="img/common/btn_edit.png" width="42" height="17" />
                                 </a>
                             </td>
               				<td>
-                            	<a href="javascript:void(0);" onClick="confirmDelete('guest_types.php?action=delete&id=<?=$row['id']?>');">
+                            	<a href="javascript:void(0);" onClick="confirmDeletePlus('guest_types.php?action=delete&id=<?=$row['id']?>&edit_id=<?=$edit_id?>',<?=$row['id']?>,'<?=$rTitle?>');">
                                 	<img src="img/common/btn_deleate.png" width="42" height="17" />
                                 </a>
                             </td>
@@ -146,9 +153,13 @@ $get = $obj->protectXSS($_GET);
 				$j++;
              	}
 			 ?>
-
       </div>
-    </div>
+		<? if($edit_id !='') {?>
+			<script>
+				$j("#boxid<?=$edit_id?>").css({backgroundColor: "#FFF0FF", color: "#990000"});
+			</script>
+		<? } ?>
+	  </div>
   </div>
 <?php
   	include_once("inc/sidebar.inc.php");
@@ -173,8 +184,10 @@ function new_table()
 }
 function cancel_new()
 {
-	$j("#name").val("");
-	$j("#insert_edit").val("0");
+	window.location = "guest_types.php";
+
+//	$j("#name").val("");
+//	$j("#insert_edit").val("0");
 }
 function check_name()
 {
@@ -199,10 +212,24 @@ function check_name()
 }
 function edit_name(id, name)
 {
-	$j("#insert_edit").val(id);
-	$j("#name").val(name);
-	$j("#title_bar").html("編集　区分");
-	$j("#new_table").fadeOut(100);
-	$j("#new_table").fadeIn(500);
+	window.location = "guest_types.php?action=edit&id="+id+"&title="+name+"&edit_id="+id;
+}
+function confirmDeletePlus(url, id, title) {
+	var edit_id="<?=$edit_id?>";
+	var msg = "";
+	if (edit_id != id) 	{
+		msg = "削除してもよろしいですか？";
+		var urls = url+"&title="+title;
+	}
+	else {
+		msg = "編集中ですが削除してもよろしいですか？";
+		var urls = url;
+	}
+
+	var agree = confirm(msg);
+	if(agree)
+	{
+		window.location = urls;
+	}
 }
 </script>
